requirements.txt
python-dotenv
streamlit
langchain-google-genai
pinecone-client
scikit-learn